requirements.txt
django<=3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.5.0
django-rest-passwordreset==1.3.0